refact: mover o conteudo dos controllers de exportação de histórico para dentro do service.
