+++ src/product/product.controller.ts
@@ -15,7 +15,6 @@ import {
 import { ProductService } from './product.service';
 import { Product } from '../models/Product';
 import { Response } from 'express';
-import { Parser } from 'json2csv';
 
 @Controller('product')
 @UsePipes(new ValidationPipe({ whitelist: true }))
@@ -79,30 +78,13 @@ export class ProductController {
   
   @Get("/:prodId/historic/export/json")
   async getHistoricJson(@Param("prodId", ParseIntPipe) id:number, @Res() res: Response){
-    const historic = await this.productService.getProductHistorics(id)
-    const jsonBuffer = Buffer.from(JSON.stringify(historic, null,2), 'utf-8')
-
-    res.set({
-      'Content-Type': 'application/json',
-      'Content-Disposition': `attachment; filename="historic-${id}.json"`,
-      'Content-Length': jsonBuffer.length,
-    });
-    return res.send(jsonBuffer)
+    return this.productService.exportHistoricJson(id, res)
   }
 
   @Get("/:prodId/historic/export/csv")
   async getHistoricCsv(@Param("prodId", ParseIntPipe) id:number, @Res() res: Response){
-    const historic = await this.productService.getProductHistorics(id)
-    const parse = new Parser({header: true})
-    const csv = parse.parse(historic)
-
-    const buffer = Buffer.from(csv, "utf-8")
-    res.set({
-      'Content-Type': 'text/csv',
-      'Content-Disposition': `attachment; filename="historic-${id}.csv"`,
-      'Content-Length': buffer.length,
-    });
-    return res.send(buffer)
+    return this.productService.exportHistoricCsv(id, res)
+    
   }
